프로젝트 생성 테스트
