--- code/Playlist_list.txt
@@ -1,4 +1,0 @@
-1. 퇴근길 플레이리스트
-2. 출근길 플레이리스트
-3. 산책 플레이리스트
-
