--- code/main.h
@@ -1,6 +1,0 @@
-#include <stdio.h>
-
-int main()
-{
-
-}
