+++ code/song_management.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34723.18
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "song_management", "song_management\song_management.vcxproj", "{EF3486DD-F2E3-46D7-9983-672AF62A05DA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EF3486DD-F2E3-46D7-9983-672AF62A05DA}.Debug|x64.ActiveCfg = Debug|x64
+		{EF3486DD-F2E3-46D7-9983-672AF62A05DA}.Debug|x64.Build.0 = Debug|x64
+		{EF3486DD-F2E3-46D7-9983-672AF62A05DA}.Debug|x86.ActiveCfg = Debug|Win32
+		{EF3486DD-F2E3-46D7-9983-672AF62A05DA}.Debug|x86.Build.0 = Debug|Win32
+		{EF3486DD-F2E3-46D7-9983-672AF62A05DA}.Release|x64.ActiveCfg = Release|x64
+		{EF3486DD-F2E3-46D7-9983-672AF62A05DA}.Release|x64.Build.0 = Release|x64
+		{EF3486DD-F2E3-46D7-9983-672AF62A05DA}.Release|x86.ActiveCfg = Release|Win32
+		{EF3486DD-F2E3-46D7-9983-672AF62A05DA}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8666E47B-2260-4313-93A6-80A1CD78A254}
+	EndGlobalSection
+EndGlobal
