+++ code/song_management/x64/Debug/song_management.log
@@ -1,0 +1,14 @@
+﻿  playlist.c
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(54,14): error C2057: 상수 식이 필요합니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(54,17): error C2466: 상수 크기 0의 배열을 할당할 수 없습니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(81,21): error C2057: 상수 식이 필요합니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(81,24): error C2466: 상수 크기 0의 배열을 할당할 수 없습니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(49,7): warning C4047: '초기화 중': 'char'의 간접 참조 수준이 'void *'과(와) 다릅니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(123,14): error C2057: 상수 식이 필요합니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(123,17): error C2466: 상수 크기 0의 배열을 할당할 수 없습니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(126,9): warning C4477: 'printf' : 서식 문자열 '%c'에 'int' 형식의 인수가 필요하지만 variadic 인수 1의 형식이 'char *'입니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(126,9): warning C4313: 'printf': 서식 문자열의 '%c'이(가) char *' 형식의 인수 1과(와) 충돌합니다.
+C:\Users\wii07\source\repos\song_management\code\song_management\playlist.c(107,12): warning C4267: '초기화 중': 'size_t'에서 'int'(으)로 변환하면서 데이터가 손실될 수 있습니다.
+  song.c
+C:\Users\wii07\source\repos\song_management\code\song_management\song.c(1,1): warning C4819: 현재 코드 페이지(949)에서 표시할 수 없는 문자가 파일에 들어 있습니다. 데이터가 손실되지 않게 하려면 해당 파일을 유니코드 형식으로 저장하십시오.
+  코드를 생성하고 있습니다...
