+++ code/song_management/x64/Debug/song_management.tlog/song_management.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.39.33519:TargetPlatformVersion=10.0.22621.0:
+Debug|x64|C:\Users\wii07\Source\Repos\song_management\code\|
